+++ client/attestation/bitVotes/bitvote.go
@@ -43,7 +43,7 @@ type WeightedBitVote struct {
 }
 
 // EncodeBitVote encodes BitVote with roundCheck
-func (b BitVote) EncodeBitVote(roundID uint64) []byte {
+func (b BitVote) EncodeBitVote(roundID uint32) []byte {
 	var encoding []byte
 	roundCheck := byte(roundID % 256)
 
@@ -59,7 +59,7 @@ func (b BitVote) EncodeBitVote(roundID uint64) []byte {
 }
 
 // EncodeBitVoteHex encodes BitVote with roundCheck prefixed with 0x to be published on chain.
-func (b BitVote) EncodeBitVoteHex(roundID uint64) string {
+func (b BitVote) EncodeBitVoteHex(roundID uint32) string {
 	str := hex.EncodeToString(b.EncodeBitVote(roundID))
 
 	return str
